# feat: Real-time token generation speedometer 🚀