# Feature: Real-time Token Generation Speedometer 🚀

## Summary

This PR adds a real-time token generation speed monitor (speedometer) to Kilo Code, displaying tokens per second (t/s) in the VS Code status bar during LLM interactions. This helps developers understand model performance and make informed decisions about model selection and optimization.

## Motivation

Currently, users have no visibility into how fast their selected models are generating responses. This information is valuable for:
- Comparing model performance in real-time
- Identifying bottlenecks in the development workflow
- Making cost/performance tradeoffs between different models
- Debugging slow responses
- Understanding the impact of context size on generation speed

## Implementation Details

### 1. Core Token Speed Monitor (`src/monitors/TokenSpeedMonitor.ts`)
- Tracks token generation speed with a sliding window approach
- Calculates current, average, and peak speeds
- Emits events for live metric updates
- Manages VS Code status bar item with configurable position

### 2. Event-Based Stream Integration (`src/api/StreamHandler.ts`)
- Integrates with existing API stream processing
- Counts tokens asynchronously using the existing tokenizer
- Event-driven architecture for loose coupling
- Handles various chunk formats (OpenAI, Anthropic, etc.)

### 3. Real-Time Speed Details View (`src/views/SpeedDetailsView.ts`)
- Interactive webview panel with live updating metrics
- Real-time chart showing speed over time
- Displays current, average, peak speeds and total tokens
- Visual status indicators for tracking/completion states

### 4. Extension Integration
- Configurable through VS Code settings
- Commands for showing details and resetting metrics
- Automatic cleanup on deactivation
- Respects VS Code UI guidelines

## User Experience

1. **Status Bar Display**: Shows real-time token generation speed with fun emoji indicators
   - 🐌 < 10 t/s (slow)
   - 🚶 10-30 t/s (walking)
   - 🏃 30-60 t/s (running)
   - 🚗 60-100 t/s (driving)
   - 🚀 > 100 t/s (rocket speed\!)

2. **Live Details Panel**: Clicking the status bar item opens a panel with:
   - Real-time updating metrics (updates every 500ms)
   - Live speed chart showing the last 60 data points
   - Current, average, and peak speeds
   - Total tokens generated and elapsed time
   - Visual status indicators

3. **Configurable**: Users can enable/disable, change position, and customize display

## Testing

The implementation has been designed to handle:
- Very fast models delivering multiple chunks < 100ms
- Metadata-only chunks with no text content
- Rapid start/stop cycles
- Network interruptions mid-stream
- Models with irregular chunk timing
- Zero-token chunks

## Breaking Changes

None - this feature is entirely additive and can be disabled via configuration.

## Configuration

New settings added:
- `kilocode.speedometer.enabled`: Enable/disable the feature (default: true)
- `kilocode.speedometer.position`: Status bar position - left or right (default: right)
- `kilocode.speedometer.showIcon`: Show emoji speed indicators (default: true)

## Screenshots

_Status bar speedometer in action:_
```
🚀 125.3 t/s
```

## Future Enhancements

- Export speed metrics to CSV/JSON
- Historical session comparison
- Model performance benchmarking mode
- Cost efficiency calculations (tokens/$ with speed consideration)
- Predictive completion time estimates
- Speed alerts/thresholds

## Checklist

- [x] Code follows the project's style guidelines
- [x] Self-review completed
- [x] Comments added for complex logic
- [x] Documentation updated
- [x] Changeset added
- [x] No console errors or warnings
- [x] Backward compatible
- [x] Respects VS Code UI guidelines
- [x] Handles all edge cases gracefully

---

Thank you for considering this feature\! The speedometer provides valuable real-time feedback that helps developers optimize their AI-assisted coding workflows.

🤖 Generated with [Claude Code](https://claude.ai/code)